refactor(shared): move contract event definitions into shared crate (#40)
